Added case insensitive dictionary, and refactored energy conversions
It is useful to be able to convert both ev and eV, without both being explicitly
added to the dictionary. I added CaseInsensitiveDict found in the `requests`
library.

I also refactored the way we actually convert the energy. The dictionary
now just holds the conversion to SI units, and then we only need to store
one number per unit we care about.