Handle url requests from python.
Using requests instead of urllib lets us ignore the weird split between
python 2 and 3. We can also use it instead of wget. AND that was where
I stole the CaseInsensitiveDict from. Win win win.

I'm using BeautifulSoup for parsing the output of requests.
That was just the only thing I knew how to use.
I also use it for the parsing in WA now.